+++ src/world/place/HangarSkin.js
@@ -19,6 +19,7 @@ const ELEMENT_KEYS = [
   'door',
   'elevator',
   'serviceBoard',
+  'hangarBayScanner',
   'crane',
   'forklift',
   'desk',